* commit '8d5f11a069033e8b5e9f059578681785184f281a':
  agregar validacion para cuando no exista registro en la tablas complementos al eliminar la solicitud de compra

+++ app/Observers/CADECO/SolicitudCompraObserver.php
@@ -61,7 +61,11 @@ class SolicitudCompraObserver extends TransaccionObserver
                 'fecha_eliminacion' => date('Y-m-d H:i:s')
             ]
         );
-        $solicitudCompra->complemento->delete();
+
+        if($solicitudCompra->complemento)
+        {
+            $solicitudCompra->complemento->delete();
+        }
     }
 
     public function deleted(SolicitudCompra $solicitudCompra)
